added helpers to core
