+++ src/components/block-editable/BlockEditable.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/core/styles';
-import * as helpers from './helpers';
+import * as helpers from '../../core/';
 import { markdownToHtml, htmlToMarkdown, filter } from '../../core/';
 
 // const whyDidYouRender = (process.env.NODE_ENV !== 'production') ?
