+++ src/components/document-translatable/DocumentTranslatable.js
@@ -1,11 +1,11 @@
-import React, {useState, useEffect} from 'react';
+import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import {makeStyles} from '@material-ui/core/styles';
+import { makeStyles } from '@material-ui/core/styles';
 import md5 from 'md5';
 
 import SectionTranslatable from '../section-translatable';
 
-import * as helpers from './helpers';
+import * as helpers from '../../core/';
 
 // const whyDidYouRender = (process.env.NODE_ENV !== 'production') ?
 //   require('@welldone-software/why-did-you-render') : undefined;
@@ -24,20 +24,20 @@ function DocumentTranslatable({
 }) {
   const classes = useStyles();
   const [sectionFocus, setSectionFocus] = useState(0);
-  const originalSections = helpers.sectionsFromMarkdown({markdown: original});
+  const originalSections = helpers.sectionsFromMarkdown({ markdown: original });
   const [translationSections, setTranslationSections] = useState([]);
-  
+
   useEffect(() => {
-    const __translationSections = helpers.sectionsFromMarkdown({markdown: translation});
+    const __translationSections = helpers.sectionsFromMarkdown({ markdown: translation });
     setTranslationSections(__translationSections);
   }, [translation]);
 
-  const setTranslationSection = ({index, translationSection}) => {
+  const setTranslationSection = ({ index, translationSection }) => {
     setSectionFocus(index);
     let _translationSections = [...translationSections];
     _translationSections[index] = translationSection;
     setTranslationSections(_translationSections);
-    const _translation = helpers.markdownFromSections({sections: _translationSections});
+    const _translation = helpers.markdownFromSections({ sections: _translationSections });
     onTranslation(_translation);
   };
 
@@ -45,7 +45,7 @@ function DocumentTranslatable({
     const key = index + md5(JSON.stringify(originalSection));
     const translationSection = translationSections[index];
     const _onTranslation = (_translationSection) =>
-      setTranslationSection({index, translationSection: _translationSection});
+      setTranslationSection({ index, translationSection: _translationSection });
     const _onSectionFocus = (expanded) => {
       if (expanded) setSectionFocus(index);
       else setSectionFocus(null);
@@ -91,7 +91,7 @@ DocumentTranslatable.propTypes = {
 DocumentTranslatable.defaultProps = {
   original: '',
   translation: '',
-  onTranslation: () => {},
+  onTranslation: () => { },
   inputFilters: [],
   outputFilters: [],
   style: {},
