+++ src/core/markdown-converter.ts
@@ -1,6 +1,7 @@
 import TurndownService from 'turndown';
 import { gfm } from 'turndown-plugin-gfm';
-import { filter } from '.';
+import { filter } from './string';
+import showdown from 'showdown';
 
 const turndownService = new TurndownService({ headingStyle: 'atx' });
 turndownService.use(gfm);
@@ -16,10 +17,20 @@ turndownService.addRule('emphasis', {
   filter: ['em'],
   replacement: (content) => `*${content}*`,
 });
+const markdownToHtmlConverter = new showdown.Converter();
+markdownToHtmlConverter.setOption('underline', true);
 
 export const htmlToMarkdown = ({ html, outputFilters = [] }) => {
   let markdown = turndownService.turndown(html);
   markdown = filter({ string: markdown, filters: outputFilters });
   if (markdown === '&#8203;') markdown = '';
   return markdown;
+};
+
+export const markdownToHtml = ({ markdown, inputFilters = [] }) => {
+  let _markdown = markdown.slice(0);
+  _markdown = filter({ string: _markdown, filters: inputFilters });
+  let html = markdownToHtmlConverter.makeHtml(_markdown);
+  if (!html || html === '') html = '<p>&#8203;</p>';
+  return html;
 };
